requirements.txt
astroid==2.3.3
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.9
django-csvimport==2.13
django-heroku==0.3.1
djangorestframework==3.11.0
flake8==4.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.21.5
PyJWT==1.7.1
pylint==2.4.4
psycopg2-binary==2.9.0
python-dotenv==0.19.2
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.11.2
